SSMP-89: Create GoogleLoginButton component - Handle logic sign up/ login using google in a hook
