+++ client/src/screens/auth/SignUp.tsx
@@ -2,11 +2,11 @@ import React from "react";
 import imageSignup from "../../assets/images/login.webp";
 import type { FormInstance } from "antd";
 import { Button, Divider, Form, Input, QRCode, message } from "antd";
-import googleIcon from "../../assets/images/devicon_google.png";
 import { usePostNewUser } from "../../hooks/user";
 import { ConfigProvider } from "antd";
 import { SignUpTheme } from "../../utils/theme";
 import { Link } from "react-router-dom";
+import GoogleLoginButton from "../../components/GoogleLoginButton";
 interface SubmitButtonProps {
   form: FormInstance;
 }
@@ -239,32 +239,7 @@ const SignUp: React.FC = () => {
             </p>
           </div>
           <Divider style={{ color: "#bdbdbd" }}>or</Divider>
-          <div className="border border-black/15 p-[20px_25px] rounded-[15px] flex items-center justify-between gap-2 cursor-pointer">
-            <div className="w-[35px] h-[35px]">
-              <img
-                src={googleIcon}
-                alt=""
-                className="w-full h-full object-cover object-center"
-              />
-            </div>
-            <div className="text-[16px] font-bold">Continue with Google</div>
-            <div>
-              <svg
-                width="28"
-                height="28"
-                viewBox="0 0 24 24"
-                fill="none"
-                xmlns="http://www.w3.org/2000/svg"
-              >
-                <path
-                  d="M10 17L15 12L10 7"
-                  stroke="grey"
-                  strokeLinecap="round"
-                  strokeLinejoin="round"
-                />
-              </svg>
-            </div>
-          </div>
+          <GoogleLoginButton />
         </div>
         <div className="w-[250px] h-[250px] absolute right-0 bottom-0 p-5 flex flex-col items-center justify-center gap-2">
           <div className="text-[13px] text-gray-400 font-normal">
